requirements.txt
pandas~=2.2.3
numpy~=2.1.3
twscrape~=0.14
google-generativeai
google-genai
pillow~=11.1.0
tweepy~=4.15.0
ace_tools
matplotlib
streamlit~=1.42.2
opencv-python~=4.11.0.86
pytesseract~=0.3.13
requests~=2.32.3
protobuf~=5.29.3
tqdm~=4.67.0
google-search-results~=2.4.2
openai
groq
